price-fetcher/requirements.txt
boto3
pynamodb
yfinance
python-dotenv
requests